requirements.txt
quart==0.18.4
websockets==6.0
requests
python-okx
asyncio